requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
pandas==2.2.2
pdfplumber==0.11.1
streamlit==1.35.0
plotly==5.22.0
python-dotenv==1.0.1
scikit-learn==1.5.0
tqdm==4.66.4
pytest==8.2.2
transformers==4.41.2
torch>=2.2.0
sentencepiece==0.2.0
